requirements.txt
dj-database-url==0.5.0
django>=2.2.10,<3.0.0
django-bootstrap3==10.0.1
django-environ==0.4.5
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework>=3.9.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2-binary==2.8.5
PyJWT==1.6.4
pytz==2018.5
raven==6.9.0
six==1.11.0
whitenoise==4.0
django-redis==4.10.0
redis==3.3.7
requests>=2.20.0
rq==1.1.0
pillow>=6.2.2
boto3==1.7.80
botocore==1.10.80
django-storages==1.6.6
python-dateutil
beautifulsoup4==4.6.3
ddtrace==0.12.1
newrelic
pre-commit==2.5.1
